requirements.txt
streamlit==1.29.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.27.6
screeninfo==0.8.1
ffmpeg-python==1.2.0
